pref（points）：积分排行榜-换用xxljob实现，赛季信息持久化+redis缓存删除

+++ tj-learning/src/main/java/com/tianji/learning/domain/vo/PointsBoardVO.java
@@ -10,7 +10,7 @@ import java.util.List;
 @ApiModel(description = "积分榜单汇总信息")
 public class PointsBoardVO {
     @ApiModelProperty("我的榜单排名")
-    private Integer rank;
+    private Long rank;
     @ApiModelProperty("我的积分值")
     private Integer points;
     @ApiModelProperty("前100名上榜人信息")
